+++ myApp/src/components/types.ts
@@ -1,3 +1,5 @@
+import { ReactNode } from "react"
+
 type Root = "E" | "A" | "D"
 type Shape = "C" | "A" | "G" | "E" | "D" | "e" | "" // lowercase e is on the D string
 type Quality = "7" | "min7" | "maj7" | "9" | "min9" | "maj9" | "11" | ""
@@ -17,4 +19,8 @@ interface Chord {
 interface CheckButton extends HTMLButtonElement {
   ariaChecked: boolean | string;
 }
-export type { Root, Shape, Quality, ShapeQuality, Extensions, Alterations, Chord, Circle, CheckButton }
+interface FretboardProps {
+  root: Root;
+  chord: Chord[];
+}
+export type { Root, Shape, Quality, ShapeQuality, Extensions, Alterations, Chord, Circle, CheckButton, FretboardProps }
