+++ myApp/src/pages/Chords.tsx
@@ -1,9 +1,186 @@
-import React from 'react';
-import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar } from '@ionic/react';
+import React, { useState } from 'react';
+import {
+  generateChord,
+  chordE7,
+  chordE9,
+  chordG7,
+  chordA7,
+  chordC7,
+  chordC9
+} from '../components/chords';
+import { getSQ, QEAMap, rootMap, shapeMap } from '../components/scales';
+import { Quality, Root, Shape, Extensions, Alterations, CheckButton } from '../components/types';
+
+import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar, IonItem, IonLabel, IonList, IonListHeader, IonSelect, IonSelectOption, IonGrid, IonRow, IonCol, IonRadioGroup, IonRadio } from '@ionic/react';
 import './Chords.css';
 import Fretboard from '../components/Fretboard';
 
 const Chords: React.FC = () => {
+  const [root, setRoot] = useState<Root>("E")
+  const [shape, setShape] = useState<Shape>("E")
+  const [quality, setQuality] = useState<Quality>("")
+  const [extensions, setExtensions] = useState<string[]>([])
+  const [alterations, setAlterations] = useState<string[]>([])
+
+  function getChord() {
+    let chord: number[][] = [[]]
+    if (root === "E") {
+      if (shape === "E" && quality.includes("7")) {
+        chord = chordE7(quality, extensions, alterations)
+      } else if (shape === "G" && quality.includes("7")) {
+        chord = chordG7(quality, extensions, alterations)
+      } else if (quality.includes("9")) {
+        chord = chordE9(quality, extensions, alterations)
+      } else {
+        chord = [[5,3]]
+      }
+    }
+    if (root === "A") {
+      if (shape === "C" && quality.includes("7")) {
+        chord = chordC7(quality, extensions, alterations)
+      } else if (shape === "A" && quality.includes("7")) {
+        chord = chordA7(quality, extensions, alterations)
+      } else if (quality.includes("9")) {
+        chord = chordC9(quality, extensions, alterations)
+      } else {
+        chord = [[4,3]]
+      }
+    }
+    if (root === "D") {
+      chord = [[3,3]]
+    }
+    return generateChord(chord, root)
+  }
+
+  function getValidExtensions() {
+    const valid = QEAMap[getSQ(shape, quality)]
+    return valid ? valid.extensions : Array<Extensions>()
+  }
+
+  function getValidAlterations() {
+    const valid = QEAMap[getSQ(shape, quality)]
+    return valid ? valid.alterations : Array<Alterations>()
+  }
+
+  function getChordDescription() {
+    const empty = "Use the below dropdowns."
+    let strQuality = ""
+    const extAndAlt = []
+
+    if (quality.includes("7") || quality.includes("9")) {
+      if (quality.includes("maj")) {
+        strQuality = "Major"
+      } else if (quality.includes("min")) {
+        strQuality = "Minor"
+      } else {
+        strQuality = "Dominant"
+      }
+      if (!alterations.includes("maj7")) extAndAlt.push("7")
+      if (alterations.includes("b5")) extAndAlt.push("b5")
+      if (alterations.includes("#5")) extAndAlt.push("#5")
+      if (quality.includes("9") || extensions.includes("9")) {
+        if (alterations.includes("b9")) extAndAlt.push("b9")
+        else if (alterations.includes("#9")) extAndAlt.push("#9")
+        else extAndAlt.push("9")
+      }
+      if (alterations.includes("#11")) extAndAlt.push("#11")
+      if (extensions.includes("13")) {
+        if (alterations.includes("b13")) extAndAlt.push("b13")
+        else extAndAlt.push("13")
+      }
+    }
+    return strQuality !== "" ? extAndAlt.length !== 0 ? 
+      `${strQuality} ${extAndAlt.reduce((total, current, i) => (total + (i<extAndAlt.length ? ", ": "") + current))}` 
+      : strQuality 
+      : empty
+  }
+
+  function setExtensionsAndClear(extensions: string[]) {
+    setAlterations([])
+    setExtensions(extensions)
+  }
+
+  function setQualityAndClear(quality: Quality) {
+    setExtensionsAndClear([])
+    setQuality(quality)
+  }
+
+  function setShapeAndClear(shape: Shape) {
+    setQualityAndClear("")
+    setShape(shape)
+  }
+
+  function setRootAndClear(root: Root) {
+    setShapeAndClear("")
+    setRoot(root)
+  }
+
+  function toggleAlterationsDisabled() {
+    if (document.querySelector(".alterations")) {
+      // The alert is up
+      const buttons = document.querySelectorAll("button.alterations-button")
+      // Disable/Activate button
+      const toggleButton = (button: Element) => {
+        if (!button.getAttribute("disabled")) {
+          button.setAttribute("disabled", "true");
+          button.classList.add("alert-checkbox-button-disabled")
+        } else {
+          button.removeAttribute("disabled")
+          button.classList.remove("alert-checkbox-button-disabled")
+        }
+      }
+      /**
+       * Calls toggleButton for other buttons depending on the current button
+       * @param value Sharp 9, Flat 5, etc. The current element's text
+       * @param current The current button
+       */
+      const checkAndToggleByValue = (value: String, current: Element) => {
+        // Special case: A7 Shape, only one alteration allowed
+        if (getSQ(shape, quality) === "A7") {
+          buttons.forEach(button => {
+            if (button !== current) {
+              toggleButton(button)
+            }
+          })
+        }
+        // Disable other 9ths
+        else if (value.includes("9")) {
+          buttons.forEach(button => {
+            if (button !== current) {
+              if (button.classList.contains("alt9")) {
+                toggleButton(button)
+              }
+            }
+          })
+        } 
+        // Disable other 5ths, #11
+        else if (value.includes("5") || value.includes("11")) {
+          buttons.forEach(button => {
+            if (button !== current) {
+              if (button.classList.contains("alt5") || button.classList.contains("alt11")) {
+                toggleButton(button)
+              }
+            }
+          })
+        }
+      }
+      
+      buttons.forEach(button => {
+        const checked = (button as CheckButton).ariaChecked
+        // On load
+        if (checked === "true") {
+          const value = ((button as Node).firstChild!.lastChild as HTMLElement).innerText
+          checkAndToggleByValue(value, button as Element)
+        }
+        // On click
+        button.addEventListener("click", (e: Event) => {
+          const value = ((e.currentTarget as Node).firstChild!.lastChild as HTMLElement).innerText
+          checkAndToggleByValue(value, e.currentTarget as Element)
+        })
+      })
+    }
+  }
+
   return (
     <IonPage>
       <IonHeader>
@@ -18,7 +195,80 @@ const Chords: React.FC = () => {
           </IonToolbar>
         </IonHeader>
         <div className="container">
-          <Fretboard />
+          <IonItem>
+            <IonLabel>Root</IonLabel>
+            <IonSelect value={root} onIonChange={e => setRootAndClear(e.detail.value)} style={{minWidth: "150px"}}>
+              {Object.keys(rootMap).map(val => <IonSelectOption value={val} key={val}>{val} String</IonSelectOption>)}
+            </IonSelect>
+          </IonItem>
+
+          <Fretboard root={root} chord={getChord()} />
+
+          <IonList>
+            <IonListHeader>
+              <IonLabel style={{textAlign: "center"}}>
+                {getChordDescription()}
+              </IonLabel>
+            </IonListHeader>
+            <IonGrid>
+              <IonRow>
+                <IonCol>
+                  <IonRadioGroup allowEmptySelection={true} value={shape} onIonChange={e => setShapeAndClear(e.detail.value)}>
+                    {shapeMap[root].map(val => (
+                      <IonItem key={val}>
+                        <IonLabel>{val.toUpperCase()} Shape</IonLabel>
+                        <IonRadio slot="end" value={val} />
+                      </IonItem>
+                    ))}
+                  </IonRadioGroup>
+                </IonCol>
+                <IonCol>
+                  <IonItem>
+                    <IonLabel>Quality</IonLabel>
+                    <IonSelect disabled={shape === ""} value={quality} onIonChange={e => setQualityAndClear(e.detail.value)} style={{minWidth: "80px"}}>
+                      <IonSelectOption value="7">{"7"}</IonSelectOption>
+                      <IonSelectOption value="maj7">{"maj7"}</IonSelectOption>
+                      <IonSelectOption value="min7">{"min7"}</IonSelectOption>
+                      <IonSelectOption value="9">{"9"}</IonSelectOption>
+                      <IonSelectOption value="maj9">{"maj9"}</IonSelectOption>
+                    </IonSelect>
+                  </IonItem>
+                  <IonItem>
+                    <IonLabel>Exts.</IonLabel>
+                    <IonSelect disabled={quality === ""} multiple={true} value={extensions} onIonChange={e => e.detail.value ? setExtensionsAndClear(e.detail.value) : null} style={{minWidth: "100px"}}>
+                      {getValidExtensions().map(val => <IonSelectOption value={val} key={val}>Add {val}</IonSelectOption>)}
+                    </IonSelect>
+                  </IonItem>
+                  <IonItem>
+                    <IonLabel>Alts.</IonLabel>
+                    <IonSelect 
+                      disabled={quality === ""} 
+                      multiple={true} 
+                      value={alterations} 
+                      style={{minWidth: "100px"}}
+                      onIonChange={e => e.detail.value ? setAlterations(e.detail.value) : null} 
+                      onIonFocus={toggleAlterationsDisabled}
+                      onIonCancel={toggleAlterationsDisabled}
+                      onIonBlur={toggleAlterationsDisabled}
+                      interfaceOptions={{
+                        cssClass: "alterations"
+                      }}
+                    >
+                      {getValidAlterations().map(val => 
+                        (!(extensions.includes("13") && getSQ(shape, quality) !== "A7") && (val.includes("5") || val.includes("11"))) ||
+                        ((extensions.includes("9") || quality.includes("9")) && !(quality.includes("min") && alterations.includes("b5")) && val.includes("9")) ||
+                        (extensions.includes("13") && val.includes("13")) ?
+                        <IonSelectOption value={val} key={val} className={`alterations-button alt${val.split(/\D/).reduce((total, current) => total+current)}`}>
+                          {val.includes("#") ? "Sharp" : "Flat"} {val.split(/\D/).reduce((total, current) => total+current)}
+                        </IonSelectOption>
+                      : null)}
+                    </IonSelect>
+                  </IonItem>
+                </IonCol>
+              </IonRow>
+            </IonGrid>
+            
+          </IonList>
         </div>
       </IonContent>
     </IonPage>
